Moved logic into a controller and fixed up my unit tests
