+++ src/scheduler_config_editor/model/lab.py
@@ -16,15 +16,18 @@ class Lab:
     @staticmethod
     def mod_lab(json_config: JsonConfig, lab: str, new_lab: str) -> None:
         """Takes in the name of an existing lab and changes it to the new lab"""
-        json_config.scheduler_config.labs[json_config.scheduler_config.labs.index(lab)] = new_lab
-        for course in json_config.scheduler_config.courses:
-            if course.lab.count(lab) == 1:
-                course.lab[course.lab.index(lab)] = new_lab
-                course.lab.sort()
-        for faculty_member in json_config.scheduler_config.faculty:
-            if lab in faculty_member.lab_preferences:
-                faculty_member.lab_preferences[new_lab] = faculty_member.lab_preferences.pop(lab)
-        json_config.scheduler_config.labs.sort()
+        try:
+            json_config.scheduler_config.labs[json_config.scheduler_config.labs.index(lab)] = new_lab
+            for course in json_config.scheduler_config.courses:
+                if course.lab.count(lab) == 1:
+                    course.lab[course.lab.index(lab)] = new_lab
+                    course.lab.sort()
+            for faculty_member in json_config.scheduler_config.faculty:
+                if lab in faculty_member.lab_preferences:
+                    faculty_member.lab_preferences[new_lab] = faculty_member.lab_preferences.pop(lab)
+            json_config.scheduler_config.labs.sort()
+        except ValueError as e:
+            raise Lab.LabMissingError(f"Lab {lab} does not exist.")
 
     @staticmethod
     def del_lab(json_config: JsonConfig, lab: str) -> None:
