+++ src/scheduler_config_editor/model/room.py
@@ -16,15 +16,18 @@ class Room:
     @staticmethod
     def mod_room(json_config: JsonConfig, room: str, new_room: str) -> None:
         """Takes in the name of an existing room and changes it to the new room"""
-        json_config.scheduler_config.rooms[json_config.scheduler_config.rooms.index(room)] = new_room
-        for course in json_config.scheduler_config.courses:
-            if course.room.count(room) == 1:
-                course.room[course.room.index(room)] = new_room
-                course.room.sort()
-        for faculty_member in json_config.scheduler_config.faculty:
-            if room in faculty_member.room_preferences:
-                faculty_member.room_preferences[new_room] = faculty_member.room_preferences.pop(room)
-        json_config.scheduler_config.rooms.sort()
+        try:
+            json_config.scheduler_config.rooms[json_config.scheduler_config.rooms.index(room)] = new_room
+            for course in json_config.scheduler_config.courses:
+                if course.room.count(room) == 1:
+                    course.room[course.room.index(room)] = new_room
+                    course.room.sort()
+            for faculty_member in json_config.scheduler_config.faculty:
+                if room in faculty_member.room_preferences:
+                    faculty_member.room_preferences[new_room] = faculty_member.room_preferences.pop(room)
+            json_config.scheduler_config.rooms.sort()
+        except ValueError as e:
+            raise Room.RoomMissingError(f"Room {room} does not exist.")
 
     @staticmethod
     def del_room(json_config: JsonConfig, room: str) -> None:
